dashboard -> area spline chart tool tips

+++ src/main/java/pk/wgu/capstone/data/service/PfmService.java
@@ -187,7 +187,7 @@ public class PfmService { // Personal Finance Management Service
         return transactionRepository.findDistinctYears(userId);
     }
 
-    public Integer getSumTransactionsByMonthAndYearAndType(Long userId, Integer year, Integer month, Type type) {
+    public BigDecimal getSumTransactionsByMonthAndYearAndType(Long userId, Integer year, Integer month, Type type) {
         return transactionRepository.getSumTransactionsByMonthAndYearAndType(userId, year, month, type);
     }
